+++ src/Parser.php
@@ -18,6 +18,7 @@ use function substr;
  * A class to parse HTTP Structured Fields from their HTTP textual representation according to RFC8941.
  *
  * Based on gapple\StructuredFields\Parser class in Structured Field Values for PHP v1.0.0.
+ *
  * @link https://github.com/gapple/structured-fields/blob/v1.0.0/src/Parser.php
  *
  * @see https://www.rfc-editor.org/rfc/rfc8941.html#section-4.2
@@ -25,7 +26,7 @@ use function substr;
  * @internal Use Dictionary::fromHttpValue(), Parameters::fromHttpValue(),
  *               OuterList::fromHttpValue(), InnerList::fromHttpValue() or Item::fromHttpValue() instead
  *
- * @phpstan-import-type DataType from Value
+ * @phpstan-import-type DataType from ValueAccess
  */
 final class Parser
 {
@@ -34,9 +35,9 @@ final class Parser
      *
      * @see https://www.rfc-editor.org/rfc/rfc8941.html#section-4.2.1
      *
-     * @return array<Value|DataType|array{
-     *     0:array<Value|DataType>,
-     *     1:array<string,Value|DataType>
+     * @return array<ValueAccess|DataType|array{
+     *     0:array<ValueAccess|DataType>,
+     *     1:array<string,ValueAccess|DataType>
      * }>
      */
     public static function parseList(Stringable|string $httpValue): array
@@ -56,9 +57,9 @@ final class Parser
      *
      * @see https://www.rfc-editor.org/rfc/rfc8941.html#section-4.2.2
      *
-     * @return array<string, Value|DataType|array{
-     *     0:array<Value|DataType>,
-     *     1:array<string,Value|DataType>
+     * @return array<string, ValueAccess|DataType|array{
+     *     0:array<ValueAccess|DataType>,
+     *     1:array<string,ValueAccess|DataType>
      * }>
      */
     public static function parseDictionary(Stringable|string $httpValue): array
@@ -84,7 +85,7 @@ final class Parser
      *
      * @see https://www.rfc-editor.org/rfc/rfc8941.html#section-4.2.1.2
      *
-     * @return array{0:array<Value|DataType>, 1:array<string,Value|DataType>}
+     * @return array{0:array<ValueAccess|DataType>, 1:array<string,ValueAccess|DataType>}
      */
     public static function parseInnerList(Stringable|string $httpValue): array
     {
@@ -141,7 +142,7 @@ final class Parser
      *
      * @see https://www.rfc-editor.org/rfc/rfc8941.html#section-4.2.1.1
      *
-     * @return array{0:array{0:array<Value|DataType>, 1:array<string,Value|DataType>}|Value, 1:int}
+     * @return array{0:array{0:array<ValueAccess|DataType>, 1:array<string,ValueAccess|DataType>}|ValueAccess, 1:int}
      */
     private static function parseItemOrInnerList(string $httpValue): array
     {
@@ -159,7 +160,7 @@ final class Parser
      *
      * @see https://www.rfc-editor.org/rfc/rfc8941.html#section-4.2.1.2
      *
-     * @return array{0:array{0:array<Value|DataType>, 1:array<string,Value|DataType>}, 1:int}
+     * @return array{0:array{0:array<ValueAccess|DataType>, 1:array<string,ValueAccess|DataType>}, 1:int}
      */
     private static function parseInnerListValue(string $httpValue): array
     {
@@ -368,7 +369,7 @@ final class Parser
     }
 
     /**
-     * @return array{0:Value, 1:string}
+     * @return array{0:ValueAccess, 1:string}
      */
     private static function parseItem(string $remainder): array
     {